Move npred cube tests to test_utils, because function is in utils.py
